rfc: Sidebar, stat components in Dashboard
