+++ test-web/src/components/Charts/DataStat.js
@@ -1,5 +1,6 @@
+//대쉬보드 현황 페이지 통계자료 표출
 import PieChart from './PieChart/pieChart';
-import StackedBarChart from './StackedBarChart';
+import StackedBarChart from './StackedBarChart/StackedBarChart';
 import Map from './choroplethMap/Map';
 import { Box } from '@mui/material';
 const DataStat = ({ startDate, endDate }) => {
